Extensibility Imporvements
This is a major overhaul that is functional on two forums live right now, but lacks the aesthetics (in ACP) to be ready for an official release. With the added functionality, developers can creates add-on modules to be placed in the modules directory and immediately recognized by Advanced Sidebox.

Many of the previous default portal boxes were ported to this modular system and have thus been made dispensable from the installation without negative consequences. Any module is optional, even the default set.

There are still a few things to work out but I am very excited about the implications of these changes. As previously noted, it had already been possible to create a plugin to add box types. I have included in this installation one sample module that uses the simple module concept (no install necessary) called Hello World(example_box). Check it out for an example of extending this plugin.

Cheers