Added Features and final extensibility tune-up.
Reconsidered some of the ideas present in the first and second drafts. Consolidated many functions and did a final sweep for language variables needed.

The modules are now single-file and a feature was added to define custom box types as templates or complete static boxes.

Interface overhaul and module management screen. Moved from the config-plugins page to it's own entry in the Config menu. Removed stylesheet as it became unnecessary.