1.2 release
Big changes in control over the sideboxes made. Admin can now control which scripts the sideboxes will be displayed on so different sets can be used for the various scripts or boxes can display on multiple or all scripts.

Added a refresh option for the theme exclude list so that when new theme are added they can be available for exclusion.

Reworked portal routines slightly to provide both sets of boxes on portal. Still not happy with it.