Initial beta commit for 1.4 feature release
At this time upgrading doesn't work fully and some side boxes may need to be edited and resaved in order to function properly. I should be able to work this out in a few days hopefully.

I am addressing Issue #21 and Issue #22 in this update.

This version is NOT READY FOR A LIVE FORUM. I can not stress that enough.