Beta update #6
Issue #41 Online Staff box incorrectly displayed any user with certain staff permissions instead of using the built-in settings. Decided to completely scrap the module and rewrote it. Much better now.

Issue #42 CodePress is now supported (if enable in ACP) for custom box editing

Several smaller updates and fine-tuning including a thorough walk-through of the HTML output and the introduction of comments for debugging.

This should signify the last of feature additions for this version and when testing has progressed enough for me to set a release date I will.