2.0.4 Bug Fixes and Enhancements
#112 Where invalid HTML was output, specifically each side box's table
was given a name property while that is not a valid attribute of an HTML
table element. Handled the job by storing the dateline in the
Ajax.SideboxPeriodicalUpdater class.

#125 where failure to globalize the language variable caused a blank
flash message to be presented when deleting addon modules.

#126 Moved templates from the global category to their own group in each
individual theme.

#127 Moved the hard-coded expander HTML to a template

#128 where modules were attempting to 'upgrade' during uninstall. Create
a constant to fal uninstall as an inappropriate time to try to upgrade
:-/

#129 Replaced archaic language ('on-line' => 'online', etc.) that had
crept into the plugin thanks to a cheap spell checker

#130 Added the option to maintain aspect ratio for avatars in the WOL

#131 replaced the title attribute with a more appropriate text

#132 where custom boxes didn't show on custom scripts or Page Manager
pages-- the cache wasn't being built properly for eval mode

#133 where CSS wasn't linked to all of the ACP pages resulting in
sub-par footer link aesthetics

Also various other changes were made that are only important to dorks
like me.