Fix to stretched avatar display
I thought we had this one covered, but it appears not. On Chrome, if "Maintain Aspect Ratio" is enabled and there are more than one avatar on a line, the avatars preceding the final avatar of the line are stretched to fit. Firefox and Internet Explorer do not do this. The "width: 100%" style on the image appears to be the culprit. Removing that style entry does not appear to change the way it displays on Firefox or Internet Explorer, even if "Maintain Aspect Ratio" is disabled.